Add language to options
