+++ src/whisper.ts
@@ -1,6 +1,7 @@
 // todo: remove all imports from file
 import { existsSync } from 'fs';
 import { DEFAULT_MODEL } from './constants';
+import { IFlagTypes } from './types';
 
 // return as syntax for whisper.cpp command
 export const createCppCommand = ({ filePath, modelName = null, modelPath = null, options = { "word_timestamps": true } }: CppCommandTypes) =>
@@ -52,6 +53,8 @@ const getFlags = (flags: IFlagTypes): string => {
   // timestamps
   if (flags["timestamp_size"]) s += " -ml " + flags["timestamp-size"];
   if (flags["word_timestamps"]) s += " -ml 1";
+  // language
+  if (flags["language"]) s += " -l " + flags["language"];
 
   return s;
 }
@@ -69,21 +72,4 @@ export const MODELS_LIST = {
   "medium.en": "ggml-medium.en.bin",
   "large-v1": "ggml-large-v1.bin",
   "large": "ggml-large.bin"
-}
-
-
-type CppCommandTypes = {
-  filePath: string,
-  modelName?: string,
-  modelPath?: string,
-  options?: IFlagTypes
-}
-
-
-export type IFlagTypes = {
-  "gen_file_txt"?: boolean,
-  "gen_file_subtitle"?: boolean,
-  "gen_file_vtt"?: boolean,
-  "timestamp_size"?: number,
-  "word_timestamps"?: boolean
 }
